requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
pillow==10.0.0
pymupdf==1.22.2
pytesseract==0.3.13
opencv-python-headless==4.7.0.72
numpy<2
python-multipart==0.0.6